+++ tests/lib/test_config.py
@@ -8,9 +8,12 @@ def test_is_env_var_set_and_not_false():
     # Not set
     assert is_env_var_set_and_not_false(ENV_VAR_NAME) == False
 
+    # Should be set by conftest
+    assert is_env_var_set_and_not_false('INVOKED_BY_PYTEST') == True
+
     # Set to empty string
     environ[ENV_VAR_NAME] = ''
-    assert is_env_var_set_and_not_false(ENV_VAR_NAME) == True
+    assert is_env_var_set_and_not_false(ENV_VAR_NAME) == False
 
     # Set to FALSE
     environ[ENV_VAR_NAME] = 'FALSE'